relevent changes

+++ GlobalDataUploder/setting/viewaddleave_setting.php
@@ -2,30 +2,37 @@
 error_reporting(0);
 MysqlConnection::connect();
 $resource = MysqlConnection::fetchAll("tbl_leavesetting");
+if (isset($_POST["btnSearch"])) {
+    $empfname = $_POST["$resultemp"];
+   
+    $sql_custom = "SELECT * FROM `tbl_leavesetting` "
+            . " WHERE `fname` LIKE '%$empfname%' ";
+            
+    $resource = MysqlConnection::fetchCustom($sql_custom);
+}
 ?>
 <div class="container-fluid">
     <div class="row-fluid">
         <div class="span12">
             <div class="widget-box">
                 <div class="widget-title"> <span class="icon"> <i class="icon-cloud"></i> </span>
-                    <h5>SEARCH LEAVE BY EMPLOYEE ID </h5>
+                    <h5>SEARCH </h5>
                 </div>
                 <div class="widget-content nopadding">
-                    <form class="form-horizontal" method="post" action="../OffsetuserServlet" name="basic_validate" id="basic_validate" novalidate="novalidate">
+                    <form class="form-horizontal" method="post" action="" name="basic_validate" id="basic_validate" novalidate="novalidate">
                         <div class="control-group">
-                            <label class="control-label ">EMPLOYEE ID :</label>
+                            <label class="control-label ">EMPLOYEE NAME :</label>
                             <div class="controls">
-                                <input type="text" name="employee_id"   autofocus="" maxlength="10" class="span11"    placeholder="EMPLOYEE ID" />
+                                <input type="text" name="empfname" value="<?php echo $resultemp["fname"] ?>"  autofocus="" maxlength="10" class="span11"    placeholder="" />
                             </div>
-                            <label class="control-label ">USERNAME :</label>
+
                             <div class="controls">
-                                <input type="text" name="username"   autofocus="" maxlength="100" class="span11"    placeholder="USERNAME" />
-                            </div>
-                            <div class="controls">
-                                <input type="hidden" name="<%= IServletConstant.ACTION%>" value="<%= IServletConstant.ACTION_SEARCH%>" />
-                                <button type="submit" class="btn btn-success">Search</button>
-                                <button type="submit" class="btn btn-danger">Clear</button>
-                                  <a href="index.php?requestPage=addleave_setting"><button type="button" class="btn btn-info">Add</button></a>
+                                <center>
+
+                                    <button type="submit" name="btnSearch" class="btn btn-success">Search</button>
+                                    <button type="submit" class="btn btn-danger">Clear</button>
+                                    <a href="index.php?requestPage=addleave_setting"><button type="button" class="btn btn-info">Add</button></a>
+                                </center>
                             </div>
                         </div>
                     </form>
@@ -40,16 +47,15 @@ $resource = MysqlConnection::fetchAll("tbl_leavesetting");
         <div class="span12">
             <div class="widget-box">
                 <div class="widget-title"> <span class="icon"><i class="icon-th"></i></span>
-                    <h5>VIEW EMPLOYEE</h5>
+                    <h5>VIEW LEAVE SETTING</h5>
                 </div>
                 <div class="widget-content nopadding">
-                    <table class="table table-bordered data-table">
+                    <table class="table table-bordered data-table" style="font-size: 11px;">
                         <thead>
                             <tr>
+                                <th style="width: 1%">#</th>
+                                <th  style="width: 1%">#</th>
 
-                                <th style="width: 1%">Edit</th>
-                                <th  style="width: 1%">Delete</th>
-                               
                                 <th>Employee Name</th>
                                 <th>Commuted Leave</th>
                                 <th>Halfpay Leave</th>
@@ -65,33 +71,30 @@ $resource = MysqlConnection::fetchAll("tbl_leavesetting");
                                 ?>
                                 <tr class="gradeX">
                                     <td>
-                                        <a href="index.php?requestPage=editleave_setting&txtId=<?php echo $result["txtId"] ?>">
+                                        <a  title="EDIT" href="index.php?requestPage=editleave_setting&txtId=<?php echo $result["txtId"] ?>">
                                             <i class="icon-pencil"></i>
                                         </a>
                                     </td>
                                     <td>
-                                        <a href="index.php?requestPage=request_delete&tblname=tbl_leavesetting&pkvalue=<?php echo $result["txtId"] ?>&location=index.php?requestPage=viewaddleave_setting">
+                                        <a title="DELETE" onclick="return confirm('Are You Sure Want to delete this Record?');" href="index.php?requestPage=request_delete&tblname=tbl_leavesetting&pkvalue=<?php echo $result["txtId"] ?>&location=index.php?requestPage=viewaddleave_setting">
                                             <i class="icon-remove"></i>
                                         </a>                                  
                                     </td>                                  
-                                    
+
                                     <td>
-                                        <?php 
-                                        $resultemp =  getEmployeeById($result["emp_id"]);
-                                                echo $resultemp["fname"]." ".$resultemp["lname"];
-                                                ?>
+                                        <?php
+                                        $resultemp = getEmployeeById($result["emp_id"]);
+                                        echo $resultemp["fname"] . " " . $resultemp["lname"];
+                                        ?>
                                     </td>
                                     <td><?php echo $result["commuted_leave"] ?></td>
                                     <td><?php echo $result["halfpay_leave"] ?></td>
                                     <td><?php echo $result["special_leave"] ?></td>
                                     <td><?php echo $result["leaving_hq"] ?></td>
-
                                 </tr>  
                                 <?php
                             }
                             ?>
-
-
                         </tbody>
                     </table>
                 </div>
@@ -100,12 +103,11 @@ $resource = MysqlConnection::fetchAll("tbl_leavesetting");
     </div>
 </div>
 
-    <?php 
-    function getEmployeeById($empId){
-        
-      $result =   MysqlConnection::fetchByPrimary("tbl_employee", $empId);
-      return($result); 
-        
-    }
+<?php
 
+function getEmployeeById($empId) {
+
+    $result = MysqlConnection::fetchByPrimary("tbl_employee", $empId);
+    return($result);
+}
 ?>
